Some refactoring to avoid circular imports
